Fixed Hydration Error + Changed Database Query
